fix: handle mouselight 'neuron' and 'neurons' data keys
