+++ src/aind_neuron_reconstruction_io/io.py
@@ -13,7 +13,6 @@ from meshparty import skeleton, meshwork
 from aind_neuron_reconstruction_io.utils import (fix_local_cloudpath, read_file, 
                                                  pull_mw_skel_colors,
                                                   get_parent_dir,get_basename)
-from aind_neuron_reconstruction_io.NeuronData import meshwork_skeleton_to_neurondata
 
 
 _cached_ccf_annotation = None
@@ -136,8 +135,8 @@ def read_json(input_file, ccf_annotate_vertices):
 
     file_content = read_file(input_file)
     data_dict = json.loads(file_content)
-    
-    neuron_list = data_dict['neurons']
+    data_key = 'neurons' if 'neurons' in data_dict else 'neuron'
+    neuron_list = data_dict[data_key] if data_key == 'neurons' else [data_dict[data_key]]
     
     # as per validation of the input file, there should be exactly one neuron in neuron_list
     neuron_dict = neuron_list[0]
@@ -366,3 +365,86 @@ def read_precomputed(project_directory, skeleton_id, ccf_annotate_vertices):
         annotate_ccf_structure(neuron_df, x_col='x', y_col='y', z_col='z')
         
     return neuron_df
+
+def meshwork_skeleton_to_neurondata(mw, ccf_annotate_vertices, get_compartment_labels):
+    """
+    Will convert a meshwork object to a NeuronData object
+
+
+    Args:
+        mw (meshparty meshworks object):
+        ccf_annotate_vertices (bool): when True will try to annotate the vertices with ccf annotations
+        get_compartment_labels (bool): when True will add compartment labels to the skeleton
+
+    Returns:
+        NeuronData:
+    """
+
+    vertices = mw.skeleton.vertices
+    n_vertices = len(vertices)
+
+    edges = mw.skeleton.edges
+    root_index = mw.skeleton._rooted.root
+
+    r_df = mw.anno.segment_properties.df[['r_eff', 'mesh_ind_filt']].set_index('mesh_ind_filt')
+    radius = r_df.loc[mw.skeleton_indices.to_mesh_region_point].r_eff.values / 1000
+
+    # get compartment labels
+    if get_compartment_labels:
+        compartment = pull_mw_skel_colors(mw, 'basal_mesh_labels', 'is_axon', 'apical_mesh_labels')
+    else:
+        compartment = np.zeros(len(radius))
+
+    # start building neuron data
+    data_list = [
+        vertices,
+        radius.reshape(-1, 1),
+        compartment.reshape(-1, 1)
+    ]
+    data_arr = np.hstack(data_list)
+
+    neuron_df = pd.DataFrame(data_arr, columns=['x', 'y', 'z', 'r', 'compartment'])
+    neuron_df['parent'] = [None] * n_vertices
+
+    for edge in edges:
+        neuron_df.loc[edge[0], 'parent'] = edge[1]
+    neuron_df.loc[root_index, 'parent'] = -2
+    neuron_df['node_id'] = neuron_df.index
+    neuron_df['node_id'] = neuron_df['node_id'] + 1
+    neuron_df['parent'] = neuron_df['parent'] + 1
+
+    # Quantify synapses by node
+    postsyn_counts_df = pd.DataFrame(mw.anno.post_syn.df['post_pt_mesh_ind_filt'].value_counts())
+    postsyn_counts_df.columns = ['count']
+    postsyn_counts_df.index.names = ['mesh']
+
+    postsyn_mesh = pd.DataFrame(mw.mesh_indices.to_skel_index_padded, columns=['skel'])
+    postsyn_mesh.index.names = ['mesh']
+    postsyn_mesh['counts'] = 0
+    postsyn_mesh.loc[postsyn_counts_df.index, 'counts'] = postsyn_counts_df['count']
+
+    presyn_counts_df = pd.DataFrame(mw.anno.pre_syn.df['pre_pt_mesh_ind_filt'].value_counts())
+    presyn_counts_df.columns = ['count']
+    presyn_counts_df.index.names = ['mesh']
+
+    presyn_mesh = pd.DataFrame(mw.mesh_indices.to_skel_index_padded, columns=['skel'])
+    presyn_mesh.index.names = ['mesh']
+    presyn_mesh['counts'] = 0
+    presyn_mesh.loc[presyn_counts_df.index, 'counts'] = presyn_counts_df['count']
+
+    postsyn_skel_counts = postsyn_mesh.groupby('skel').sum().to_dict()['counts']
+    presyn_skel_counts = presyn_mesh.groupby('skel').sum().to_dict()['counts']
+
+    neuron_df['postsynaptic_count'] = neuron_df['node_id'].map(postsyn_skel_counts)
+    neuron_df['presynaptic_count'] = neuron_df['node_id'].map(presyn_skel_counts)
+    if ccf_annotate_vertices:
+        io.annotate_ccf_structure(neuron_df, x_col='x', y_col='y', z_col='z')
+    else:
+        neuron_df['allenId'] = [0] * len(neuron_df)
+
+    neuron_df = neuron_df[
+        ['node_id', 'compartment', 'x', 'y', 'z', 'r', 'parent', 'allenId', 'presynaptic_count', 'postsynaptic_count']]
+
+    neuron_data = NeuronData(input_data=neuron_df, ccf_annotate_vertices=False)
+
+    return neuron_data
